requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
httpie==2.4.0
Pygments==2.9.0
pytest-django==4.4.0
requests==2.25.1
